executed cleanup wiz & format
git-svn-id: file:///home/igor/dev/stuff/svnbackup/trunk@851 ef7698a4-5110-0410-9fc6-c7eb3693863f

+++ wicket-contrib-data/src/java/wicket/contrib/data/model/bind/OrderByLink.java
@@ -118,7 +118,7 @@ public class OrderByLink extends Link<Integer>
 	 */
 	private void switchState()
 	{
-		Integer state = (Integer) getModelObject();
+		Integer state = getModelObject();
 
 		if (state.equals(NONE) || state.equals(UP))
 		{
@@ -151,7 +151,7 @@ public class OrderByLink extends Link<Integer>
 		 */
 		public String getObject()
 		{
-			Integer state = (Integer) OrderByLink.this.getModelObject();
+			Integer state = OrderByLink.this.getModelObject();
 			if (state.equals(UP))
 			{
 				return "wicket_orderUp";
